+++ timezones/tasks.py
@@ -1,7 +1,0 @@
-"""
-our tasks
-"""
-
-import logging
-
-logger = logging.getLogger(__name__)